requirements.txt
flask
pandas 
supabase 
requests

transformers 
torch 
sentencepiece 
python-dotenv